Add hook 'Component.before_rendering'
Sometimes we want to compute something to decide if the component should be
destroyed or not, however putting this kind of computation directly in
__init__ doesn't play well with caching (unless you entangle your __init__
with caches as well).

This new hook 'before_render' allows to defer the computation until the
render-time.

NB: We're preparing a subclass CachedHTMXComponent which leverages allows to
integrate with caches.