Rewrite signals/event handling.
Events are now placed in the Repository, side by side with signals.  Events
are a global property of the entire HTTP request cycle.  The cannot be
attached to a single component.

Even in the case of a component destroying itself after some event, those
events should be triggered.

The cycle of signals/events is capped to 50 repetitions.  We assume anything
more to be an infinite cycle.  This cycle is controlled by the consumption and
detection of signals and events (aka a deque of signals and events).

Components are awaken before trying to read them.