Fix 'Rewrite signals/event handling'.
This fixes and continues commit 0c2db711646b0578b2323dd4de5f473450a8d2d7.

We could be rendering the main component more than once, if it was subscribed
to an event/signal it itself just emitted.

A simple pattern like:

```
class Tooltip(PydanticComponent):
    def _event_handler(self, event: SetTooltip):
        self.tooltip = event.tooltip

    def close(self):
        self.controller.emit(SetTooltip(None))

@dataclass
class SetTooltip:
    tooltip: str | None

```

Would cause 'close' to be render the component (without the updated 'tooltip')
and then after calling the event handler.

Similar things could happen with triggers.

This commit fixes this by detecting if the _main component_ (the one that
received the HTTP request) was updated during the dispatching of signals, and
avoid to render it twice in that case.